WS Proposto adicionei restricoes no Domain

+++ WS_Proposta/proposta/src/main/java/wsproposta/proposta/DTO/assemblers/PropostaDomainDTOAssembler.java
@@ -8,7 +8,7 @@ public class PropostaDomainDTOAssembler {
 
     private PropostaDomainDTOAssembler(){}
 
-    public PropostaDTO toDTO (int codUtilizador, int nifOrganizacao, int codEdicao, String sTitulo, String sProblema, String sObjetivo, String estado){
+    public PropostaDTO toDTO (int codUtilizador, long nifOrganizacao, int codEdicao, String sTitulo, String sProblema, String sObjetivo, String estado){
         return new PropostaDTO(codUtilizador, nifOrganizacao, codEdicao, sTitulo, sProblema, sObjetivo, estado);
     }
 }
